refactor: seperate svg to components
