+++ client/src/components/user-profile/background.jsx
@@ -1,0 +1,56 @@
+import React from "react";
+
+export default function Background() {
+  return (
+    <div>
+      <div className="absolute right-0 top-52 -z-20">
+        <svg
+          width="61"
+          height="74"
+          viewBox="0 0 61 74"
+          fill="none"
+          xmlns="http://www.w3.org/2000/svg"
+        >
+          <circle cx="37" cy="37" r="37" fill="#C6DCFF" />
+        </svg>
+      </div>
+      <div className="absolute right-[-1rem] top-28 md:right-[9rem] md:top-[9rem] -z-20">
+        <svg
+          width="51"
+          height="51"
+          viewBox="0 0 51 51"
+          fill="none"
+          xmlns="http://www.w3.org/2000/svg"
+        >
+          <path
+            d="M11.3581 19.9099L37.1499 15.9774L27.6597 40.28L11.3581 19.9099Z"
+            stroke="#FBAA1C"
+            strokeWidth="3"
+          />
+        </svg>
+      </div>
+      <div className="absolute left-[-15px] top-28 md:left-[2rem] md:top-[12rem] -z-20">
+        <svg
+          width="27"
+          height="27"
+          viewBox="0 0 27 27"
+          fill="none"
+          xmlns="http://www.w3.org/2000/svg"
+        >
+          <circle cx="13.1741" cy="13.1741" r="13.1741" fill="#C6DCFF" />
+        </svg>
+      </div>
+      <div className="absolute left-6 top-20 md:left-[6rem] md:top-[8rem] -z-20">
+        <svg
+          width="11"
+          height="11"
+          viewBox="0 0 11 11"
+          fill="none"
+          xmlns="http://www.w3.org/2000/svg"
+        >
+          <circle cx="5.5" cy="5.5" r="4" stroke="#2F5FAC" strokeWidth="3" />
+        </svg>
+      </div>
+    </div>
+  );
+}
